Fix tests to account for returned logic
