+++ packages/perseus/src/widgets/number-line/validate-number-line.test.ts
@@ -22,23 +22,6 @@ describe("validateNumberLine", () => {
         );
     });
 
-    it("is invalid when end state is the same as beginning state", () => {
-        // Arrange
-        const userInput: PerseusNumberLineUserInput = {
-            isTickCrtl: true,
-            rel: "eq",
-            numDivisions: 10,
-            divisionRange: [-10, 10],
-            numLinePosition: 0,
-        };
-
-        // Act
-        const validationError = validateNumberLine(userInput);
-
-        // Assert
-        expect(validationError).toHaveInvalidInput();
-    });
-
     it("returns null when validation passes", () => {
         // Arrange
         const userInput: PerseusNumberLineUserInput = {
